reduce number of modifiers
