--- packages/expo-notifications/android/src/main/java/expo/modules/notifications/displayers/modifiers/AppIdModifier.java
@@ -1,16 +1,0 @@
-package expo.modules.notifications.displayers.modifiers;
-
-import android.content.Context;
-import android.os.Bundle;
-import androidx.core.app.NotificationCompat;
-
-import static expo.modules.notifications.NotificationConstants.NOTIFICATION_APP_ID_KEY;
-
-public class AppIdModifier implements NotificationModifier {
-  @Override
-  public void modify(NotificationCompat.Builder builder, Bundle notification, Context context, String appId) {
-    if (!notification.containsKey(NOTIFICATION_APP_ID_KEY)) {
-      notification.putString(NOTIFICATION_APP_ID_KEY, appId);
-    }
-  }
-}
