--- packages/expo-notifications/android/src/main/java/expo/modules/notifications/displayers/modifiers/ColorModifier.java
@@ -1,17 +1,0 @@
-package expo.modules.notifications.displayers.modifiers;
-
-import android.content.Context;
-import android.os.Bundle;
-import androidx.core.app.NotificationCompat;
-
-import static expo.modules.notifications.NotificationConstants.NOTIFICATION_COLOR;
-
-public class ColorModifier implements NotificationModifier {
-
-  @Override
-  public void modify(NotificationCompat.Builder builder, Bundle notification, Context context, String appId) {
-    if (notification.containsKey(NOTIFICATION_COLOR) && notification.getString(NOTIFICATION_COLOR) != null) {
-      builder.setColor(notification.getInt(NOTIFICATION_COLOR));
-    }
-  }
-}
