--- packages/expo-notifications/android/src/main/java/expo/modules/notifications/displayers/modifiers/StickyModifier.java
@@ -1,19 +1,0 @@
-package expo.modules.notifications.displayers.modifiers;
-
-import android.content.Context;
-import android.os.Bundle;
-import androidx.core.app.NotificationCompat;
-
-import static expo.modules.notifications.NotificationConstants.NOTIFICATION_STICKY;
-
-public class StickyModifier implements NotificationModifier {
-  @Override
-  public void modify(NotificationCompat.Builder builder, Bundle notification, Context context, String appId) {
-    if (!notification.getBoolean(NOTIFICATION_STICKY)) {
-      builder.setAutoCancel(true);
-    } else {
-      builder.setAutoCancel(false);
-      builder.setOngoing(true);
-    }
-  }
-}
