--- packages/expo-notifications/android/src/main/java/expo/modules/notifications/displayers/modifiers/TitleModifier.java
@@ -1,16 +1,0 @@
-package expo.modules.notifications.displayers.modifiers;
-
-import android.content.Context;
-import android.os.Bundle;
-import androidx.core.app.NotificationCompat;
-
-import static expo.modules.notifications.NotificationConstants.NOTIFICATION_TITLE;
-
-public class TitleModifier implements NotificationModifier {
-  @Override
-  public void modify(NotificationCompat.Builder builder, Bundle notification, Context context, String appId) {
-    if (notification.containsKey(NOTIFICATION_TITLE)) {
-      builder.setContentTitle(notification.getString(NOTIFICATION_TITLE));
-    }
-  }
-}
